+++ changelog.rst
@@ -3,6 +3,9 @@
 Change Log
 ##########
 
+- 2022-04-29
+   - Define a color for ``swiper-line-face`` which caused comments to be unreadable.
+
 - 2022-04-28
   - Use a slightly brighter background color for ``whitespace-tab``.